DIGG-452: Updating search result list to show the publisher in the current language
