+++ providers/EntrystoreProvider/index.tsx
@@ -3,6 +3,7 @@ import { useRouter } from "next/router";
 import React, { createContext, useEffect, useState } from "react";
 import { EnvSettings } from "@/env/EnvSettings";
 import { SettingsUtil } from "@/env/SettingsUtil";
+import { getLocalizedValue } from "@/utilities";
 
 //unfortunate hack to get a entrystore class instance, script is inserted in head
 declare var ESJS: any;
@@ -96,90 +97,6 @@ export const EntrystoreProvider: React.FC<EntrystoreProviderProps> = ({
         },
       );
     }
-  };
-
-  /**
-   * Search graph for localized value from meta graph
-   *
-   * Supports uri-types (will fetch uri and display foaf:name, if any)
-   * TODO: support
-   *
-   * value type retrieve order:
-   * 1. exists in sent in lang
-   * 2. exists in fallback lang (en)
-   * 3. take first
-   *
-   * @param metadataGraph
-   * @param prop
-   * @param lang
-   */
-  const getLocalizedValue = async (
-    metadataGraph: any,
-    prop: any,
-    lang: string,
-    es: any,
-    options: { uriTypeName?: string; resourceURI?: string } = {
-      uriTypeName: "foaf:name",
-    },
-  ) => {
-    let val = "";
-    const fallbackLang = "en";
-    const { uriTypeName, resourceURI } = options;
-
-    const stmts = metadataGraph.find(resourceURI, prop);
-    if (stmts.length > 0) {
-      const obj: any = {};
-      for (let s = 0; s < stmts.length; s++) {
-        let stType = stmts[s].getType();
-        let stValue = stmts[s].getValue();
-
-        if (stType && stType == "uri" && !stValue.includes("mailto:")) {
-          let res = await resourcesSearch([stValue], es);
-          if (res && res.length > 0) {
-            let meta = res[0].getAllMetadata();
-
-            if (meta)
-              obj[stmts[s].getLanguage() || ""] = getLocalizedValue(
-                meta,
-                uriTypeName || "foaf:name",
-                nextLang,
-                es,
-                { resourceURI },
-              );
-          } else obj[stmts[s].getLanguage() || ""] = stValue;
-        } else obj[stmts[s].getLanguage() || ""] = stValue;
-      }
-
-      if (typeof obj[lang] != "undefined") {
-        val = obj[lang];
-      } else if (obj[fallbackLang] && fallbackLang != lang) {
-        val = obj[fallbackLang];
-      } else {
-        val = Object.entries(obj)[0][1] as string;
-      }
-    }
-
-    return val;
-  };
-
-  /**
-   * Make SolrSearch and retrive entries from entryscape
-   * Does not handle to large resource arrays, can leed to request URI errors,
-   * use in batches
-   *
-   * @param resources
-   * @param es
-   */
-  const resourcesSearch = (resources: string[], es: any): Promise<any> => {
-    return new Promise<any>((resolve) => {
-      let esQuery = es.newSolrQuery();
-      esQuery
-        .resource(resources, null)
-        .getEntries(0)
-        .then((children: any) => {
-          resolve(children);
-        });
-    });
   };
 
   const parseEmail = (mailStr: string) => {
